+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Fixed
+- **Uptime Kuma** - Self-hosted uptime monitoring with 90+ notification services
+
 ## [1.3.3] - 2026-02-27
 
 ### Fixed